job-scout/requirements.txt
openai>=1.0.0
apify-client>=1.6.0
notion-client>=2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pdfplumber>=0.10.0
numpy>=1.24.0
flashrank>=0.2.10
ddgs>=7.0.0
